PhyloPOMP.jl

For doing phylodynamics in Julia. PhyloPOMP.jl enables computation of the phylodynamic likelihood using tools from POMP.jl.

Function Documentation

PhyloPOMP.@demes — Macro

@demes Name deme1 deme2 ...

Creates a new module, Name, which contains an enumeration of the demes deme1, deme2, ....

PhyloPOMP.parse_newick — Function

parse_newick(input; demes, t0, time)

Parse the Newick-format string (or vector of strings) input. It is possible to include metadata tags of format [&&PhyloPOMP ...] in the label-strings. All tips are assumed to be samples. Zero-length branches are dropped.

Arguments:

demes is a Module enumerating the demes (see @demes).

t0 is the assumed root-time.

time is the (optional) final-time.

PhyloPOMP.newick — Function

newick(g; extended = true, sigdigits = 6)

Generates a String containing a Newick-format description of the Genealogy g. If extended = true (the default), then the PhyloPOMP-flavored Newick extension is used (i.e., metadata tags are included).

PhyloPOMP.guide — Function

guide(g, m)

g is a Genealogy.

m: is an FSMarkovProc encoding the guiding finite-state Markov process. Construct it with a call to fsmarkov.

PhyloPOMP.fsmarkov — Function

fsmarkov(...)

Constructs a finite-state Markov process in continuous time, represented as an object of type FSMarkovProc. Each argument is of the form D => prob or (D1,D2) => cond, where D, D1, D2 are demes, prob is a probability, and cond is a conductance. Specifically,

D => prob indicates that the stationary distribution on deme D is proportional to prob.

(D1,D2) => cond mandates that the conductance between D1 and D2 is cond.

Reference

PhyloPOMP.PhyloPOMP — Module

PhyloPOMP

Phylodynamic inference based on POMP models (partially observed Markov processes).

Documentation for PhyloPOMP.jl v0.0.7-1.

PhyloPOMP.demekron — Method

demekron([F::Type], d)

Kronecker delta for deme d.

PhyloPOMP.fsmarkov — Method

fsmarkov(...)

Constructs a finite-state Markov process in continuous time, represented as an object of type FSMarkovProc. Each argument is of the form D => prob or (D1,D2) => cond, where D, D1, D2 are demes, prob is a probability, and cond is a conductance. Specifically,

D => prob indicates that the stationary distribution on deme D is proportional to prob.

(D1,D2) => cond mandates that the conductance between D1 and D2 is cond.

PhyloPOMP.guide — Method

guide(g, m)

g is a Genealogy.

m: is an FSMarkovProc encoding the guiding finite-state Markov process. Construct it with a call to fsmarkov.

PhyloPOMP.newick — Method

newick(g; extended = true, sigdigits = 6)

Generates a String containing a Newick-format description of the Genealogy g. If extended = true (the default), then the PhyloPOMP-flavored Newick extension is used (i.e., metadata tags are included).

PhyloPOMP.parse_newick — Method

parse_newick(input; demes, t0, time)

Parse the Newick-format string (or vector of strings) input. It is possible to include metadata tags of format [&&PhyloPOMP ...] in the label-strings. All tips are assumed to be samples. Zero-length branches are dropped.

Arguments:

demes is a Module enumerating the demes (see @demes).

t0 is the assumed root-time.

time is the (optional) final-time.

PhyloPOMP.relhaz — Method

relhaz(g, t, n)

Compute the relative reverse-time hazards associated with filter guide g at time t relative to the target at guide-node n.

PhyloPOMP.repair! — Method

repair!(G)

Re-sorts the genealogical nodes. Renames all nodes and corrects the parent/child relationships. Traces sample-lineages.

PhyloPOMP.@demes — Macro

@demes Name deme1 deme2 ...

Creates a new module, Name, which contains an enumeration of the demes deme1, deme2, ....

PhyloPOMP.@marks — Macro

@marks Name mark1 mark2 ...

Creates a new module, Name, which contains an enumeration of the jump-marks mark1, mark2, ....

PhyloPOMP.FSMarkovProc — Type

FSMarkovProc

Represents a finite-state Markov process in continuous time.

PhyloPOMP.GenealNode — Type

GenealNode{D}

Implements a genealogical node. The module D enumerates the demes (see @demes).

PhyloPOMP.Genealogy — Type

Genealogy{D}

Implements a genealogy over a time interval [t0,t]. The module D enumerates the demes (see @demes). Internally, this is represented as a time-ordered sequence of genealogical nodes (represented by GenealNode objects).

PhyloPOMP.Guide — Type

Guide{F,D}

This type holds a "filter-guide" based on a reverse-time finite-state Markov process. Construct it with a call to guide. F is an AbstractFloat type and D is the enumeration of the demes. It contains the FSMarkovProc object encoding the Markov process and a sequence of GuideNode objects.

PhyloPOMP.GuideNode — Type

GuideNode{F,D}

A GuideNode contains the information needed to guide a phylodynamic Monte Carlo filter on a particular interval between genealogical events. F is an AbstractFloat type and D is the enumeration of the demes (see @demes).

A GuideNode contains:

the time interval to which it pertains

the type of event (NodeType) at the left endpoint of the interval

the name of the parent lineage (ancestral to the genealogical node at the left endpoint)

the names of the child lineages (immediately descended from the parent node)

the names of all lineages present on the interval

the guide probability for each child lineage at the left interval endpoint

the target probabilities for all lineages at the right interval endpoint.

PhyloPOMP.NodeType — Type

NodeType

Genealogical nodes (see GenealNode) can be one of three types: Root, Node, or Sample.

Index

PhyloPOMP.PhyloPOMP

PhyloPOMP.FSMarkovProc

PhyloPOMP.GenealNode

PhyloPOMP.Genealogy

PhyloPOMP.Guide

PhyloPOMP.GuideNode

PhyloPOMP.NodeType

PhyloPOMP.demekron

PhyloPOMP.fsmarkov

PhyloPOMP.fsmarkov

PhyloPOMP.guide

PhyloPOMP.guide

PhyloPOMP.newick

PhyloPOMP.newick

PhyloPOMP.parse_newick

PhyloPOMP.parse_newick

PhyloPOMP.relhaz

PhyloPOMP.repair!

PhyloPOMP.@demes

PhyloPOMP.@demes

PhyloPOMP.@marks