Reference

PhyloPOMP.PhyloPOMP — Module

PhyloPOMP

Phylodynamic inference based on POMP models (partially observed Markov processes).

Documentation for PhyloPOMP.jl v0.0.10-3.

PhyloPOMP.chop! — Method

chop!(y, i, b, j, c)

Chop lineage b from deme i; add lineage c in deme j.

PhyloPOMP.chop! — Method

chop!(y, i, b)

Chop lineage b from deme i.

PhyloPOMP.demekron! — Method

demekron!(v, d)

Kronecker delta for deme d. Replaces the entries of v with zeros in all places except the one corresponding to deme d.

PhyloPOMP.demekron — Method

demekron([F::Type], d)

Kronecker delta for deme d. By default, F = Float64.

PhyloPOMP.enum2name — Method

enum2name(demeset)

Returns a Dict that maps demes to strings.

PhyloPOMP.fork! — Method

fork!(y, i, b, j, c)

Fork lineage b in deme i to lineages given in the vector c in demes given respectively by the vector j.

PhyloPOMP.fsmarkov — Method

fsmarkov(...)

Constructs a finite-state Markov process in continuous time, represented as an object of type FSMarkovProc. Each argument is of the form D => prob or (D1,D2) => cond, where D, D1, D2 are demes, prob is a probability, and cond is a conductance. Specifically,

D => prob indicates that the stationary distribution on deme D is proportional to prob.

(D1,D2) => cond mandates that the conductance between D1 and D2 is cond.

PhyloPOMP.guide — Method

guide(g, m, knowledge!)

g is a Genealogy.

m: is an FSMarkovProc encoding the guiding finite-state Markov process. Construct it with a call to fsmarkov.

knowledge! is a function with signature knowledge!(v; deme, type, time). It should overwrite the vector v with probabilities of being in each possible deme, according to the given deme, type, and time. By default, knowledge! = known_deme! (see known_deme!).

PhyloPOMP.known_deme! — Method

known_deme!(v; deme, type, time)

The default behavior of guide is to assume the deme is known if and only if it is fixed in the genealogy. This function implements that default.

PhyloPOMP.name2enum — Method

name2enum(demes)

Returns a function that maps strings to demes.

PhyloPOMP.newick — Method

newick(g; extended = true, sigdigits = 6)

Generates a String containing a Newick-format description of the Genealogy g. If extended = true (the default), then the PhyloPOMP-flavored Newick extension is used (i.e., metadata tags are included).

PhyloPOMP.parse_newick — Method

parse_newick(input; demes, t0, time)

Parse the Newick-format string (or vector of strings) input. It is possible to include metadata tags of format [&&PhyloPOMP ...] in the label-strings. All tips are assumed to be samples. Zero-length branches are dropped.

Arguments:

demes is a Module enumerating the demes (see @demes).

t0 is the assumed root-time.

time is the (optional) final-time.

PhyloPOMP.plant! — Method

plant!(y, i, b)

Plant lineage b in deme i within coloring y.

PhyloPOMP.rcateg — Function

rcateg(p, s, prob = false)

This call returns a draw from the categorical distribution on the set s. The latter may be a Set or a BitSet.

PhyloPOMP.rcateg — Function

rcateg(p, v, prob = false)

This call returns a draw from the categorical distribution on the vector v.

PhyloPOMP.rcateg — Function

rcateg(p, prob = false)

If p is a vector of weights, rcateg(p) returns a draw from the categorical distribution on 1:length(p) with these weights. It also returns the sum of the weights.

If prob=true, the normalized weight of the selected category is returned as a third value.

It is not necessary for the weights to be normalized: this is accomplished internally.

PhyloPOMP.rcateg — Method

rcateg(p, e, prob = false)

This call returns a draw from the categorical distribution on the enumeration e. e should be an enumeration-type created, e.g., by @demes.

PhyloPOMP.relhaz — Method

relhaz(t, g, n, i, j, cols)

In this form, cols is a Coloring.

PhyloPOMP.relhaz — Method

relhaz(t, g, n, i, j, lins)

Compute the relative reverse-time hazards associated with filter guide g at time t relative to the target at guide-node n. Only columns denoted by the indices in lins are considered.

This returns a vector of hazards for the i ⟶ j transition relative to the hazard at stationarity.

PhyloPOMP.repair! — Method

repair!(G)

Re-sorts the genealogical nodes. Renames all nodes and corrects the parent/child relationships. Traces sample-lineages.

PhyloPOMP.swap! — Method

swap!(y, i, j, b)

Swap lineage b from deme i to deme j in coloring y.

PhyloPOMP.@demes — Macro

@demes Name deme1 deme2 ...

Creates a new demeset, Name, which contains an enumeration of the demes deme1, deme2, .... A demeset is a Module, which contains the enumeration and the enumeration type.

PhyloPOMP.@indicator — Macro

@indicator condition value type

Returns value if condition=true, zero(type) otherwise. By default, type = Float64.

PhyloPOMP.@marks — Macro

@marks Name mark1 mark2 ...

Creates a new module, Name, which contains an enumeration of the jump-marks mark1, mark2, ....

PhyloPOMP.Coloring — Type

    Coloring(d)

The Coloring struct holds information on the local coloring of a genealogy. Construct it with the call Coloring(d), where d is a demeset (see @demes).

PhyloPOMP.FSMarkovProc — Type

FSMarkovProc

Represents a finite-state Markov process in continuous time.

PhyloPOMP.GenealNode — Type

GenealNode{E}

Implements a genealogical node. The type E is an enumeration of the demes (see @demes).

PhyloPOMP.Genealogy — Type

Genealogy{D}

Implements a genealogy over a time interval [t0,t]. The module D enumerates the demes and should be constructed via a call to @demes. Internally, a genealogy is represented as a time-ordered sequence of genealogical nodes (represented by GenealNode objects).

PhyloPOMP.Guide — Type

Guide{F,D}

This type holds a "filter-guide" based on a reverse-time finite-state Markov process. Construct it with a call to guide. F is an AbstractFloat type and D is the enumeration of the demes. It contains the FSMarkovProc object encoding the Markov process and a sequence of GuideNode objects.

PhyloPOMP.GuideNode — Type

GuideNode{F,D}

A GuideNode contains the information needed to guide a phylodynamic Monte Carlo filter on a particular interval between genealogical events. F is an AbstractFloat type and D is the enumeration of the demes (see @demes).

A GuideNode contains:

the time interval to which it pertains

the type of event (NodeType) at the left endpoint of the interval

the name of the parent lineage (ancestral to the genealogical node at the left endpoint)

the names of the child lineages (immediately descended from the parent node)

the names of all lineages present on the interval

the guide probability for each child lineage at the left interval endpoint

the target probabilities for all lineages at the right interval endpoint.

PhyloPOMP.NodeType — Type

NodeType

Genealogical nodes (see GenealNode) can be one of three types: Root, Node, or Sample.

PhyloPOMP.Unstructured — Module

Unstructured

A demeset for the unstructured (1-deme) case.