SEIR model

Naïve filter

PhyloPOMP.NaiveSEIR — Module

NaiveSEIR

A module containing an implementation of the phylodynamic filter for an SEIR model, using so-called naïve proposals. These proposals are non-anticipatory. In particular, population events occurring at any time result in color changes with a probability proportional to the fraction of appropriate lineages represented in the genealogy at that time.

PhyloPOMP.NaiveSEIR.filter_pomp — Method

filter_pomp(gen; β = 4.0, σ = 1.0, γ = 1.0, ω = 1.0, ψ = 0.02, χ = 0.0,
     pop = 100, S0 = 0.9, E0 = 0.0, I0 = 0.02, R0 = 0.08)

Constructs a pomp object based on the genealogy gen and implementing the naïve genealogical filter.

PhyloPOMP.NaiveSEIR.seir_trees — Constant

seir_trees

Several trees simulated from an SEIR model, for use in testing the phylodynamic filter methods. These are given as a vector of Newick-format strings.

An example constructing the pomp object and running a particle filter.

using Random: seed!
using PhyloPOMP
using PhyloPOMP.NaiveSEIR

seed!(351956486)

g = parse_newick(NaiveSEIR.seir_trees[1], time = 50.0)
p = NaiveSEIR.filter_pomp(g)
pf = pfilter(p, Np = 1000)
round(logLik(pf),digits=1)

-218.1

Guided filter

PhyloPOMP.GuidedSEIR — Module

GuidedSEIR

A module containing an implementation of the phylodynamic filter for an SEIR model, using a filter guide and "soft" proposals. That is, when population-process events occur, the guide can steer those events preferentially onto (or away from) particular branches, but the overall event rate remains equal to that of the underlying population process.

PhyloPOMP.GuidedSEIR.filter_pomp — Method

filter_pomp(g; β = 4.0, σ = 1.0, γ = 1.0, ω = 1.0, ψ = 0.02, χ = 0.0,
     pop = 100, S0 = 0.9, E0 = 0.0, I0 = 0.02, R0 = 0.08)

Constructs a pomp object based on the filter guide g.

PhyloPOMP.GuidedSEIR.knowledge! — Method

knowledge!(v; deme, type, time)

This function should return true if the guide probabilities will be fixed at this node and false otherwise. If the former, it should fill the vector v with an appropriate probability vector.

PhyloPOMP.GuidedSEIR.seir_trees — Constant

seir_trees

Several trees simulated from an SEIR model, for use in testing the phylodynamic filter methods. These are given as a vector of Newick-format strings.

An example constructing the pomp object and running a particle filter.

using Random: seed!
using PhyloPOMP
using PhyloPOMP.GuidedSEIR
using PhyloPOMP.GuidedSEIR.Demes: Expos, Infec

seed!(351956486)

g = parse_newick(GuidedSEIR.seir_trees[1], time = 50.0)
p = GuidedSEIR.filter_pomp(g,fsmarkov(Expos=>0.1,Infec=>1,(Expos,Infec)=>1))
pf = pfilter(p, Np = 1000)
round(logLik(pf),digits=1)

-216.7

Hard-guided filter

PhyloPOMP.HardSEIR — Module

HardSEIR

A module containing an implementation of the phylodynamic filter for an SEIR model, using a filter guide and "hard" proposals. That is, color changes can be proposed on branches at rates that exceed the event rates in the underlying population process.

PhyloPOMP.HardSEIR.filter_pomp — Method

filter_pomp(g; β = 4.0, σ = 1.0, γ = 1.0, ω = 1.0, ψ = 0.02, χ = 0.0,
     pop = 100, S0 = 0.9, E0 = 0.0, I0 = 0.02, R0 = 0.08)

Constructs a pomp object based on the filter guide g.

PhyloPOMP.HardSEIR.knowledge! — Method

knowledge!(v; deme, type, time)

This function should return true if the guide probabilities will be fixed at this node and false otherwise. If the former, it should fill the vector v with an appropriate probability vector.

PhyloPOMP.HardSEIR.seir_trees — Constant

seir_trees

Several trees simulated from an SEIR model, for use in testing the phylodynamic filter methods. These are given as a vector of Newick-format strings.

An example constructing the pomp object and running a particle filter.

using Random: seed!
using PhyloPOMP
using PhyloPOMP.HardSEIR
using PhyloPOMP.HardSEIR.Demes: Expos, Infec

seed!(351956486)

g = parse_newick(HardSEIR.seir_trees[1], time = 50.0)
p = HardSEIR.filter_pomp(g,fsmarkov(Expos=>0.1,Infec=>1,(Expos,Infec)=>1))
pf = pfilter(p, Np = 1000)
round(logLik(pf),digits=1)

-215.7