PhyloPOMP.jl

For doing phylodynamics in Julia. PhyloPOMP.jl enables computation of the phylodynamic likelihood using tools from POMP.jl.

Function Documentation

PhyloPOMP.@demes — Macro

@demes Name deme1 deme2 ...

Creates a new demeset, Name, which contains an enumeration of the demes deme1, deme2, .... A demeset is a Module, which contains the enumeration and the enumeration type.

PhyloPOMP.parse_newick — Function

parse_newick(input; demes, t0, time)

Parse the Newick-format string (or vector of strings) input. It is possible to include metadata tags of format [&&PhyloPOMP ...] in the label-strings. All tips are assumed to be samples. Zero-length branches are dropped.

Arguments:

demes is a Module enumerating the demes (see @demes).

t0 is the assumed root-time.

time is the (optional) final-time.

PhyloPOMP.newick — Function

newick(g; extended = true, sigdigits = 6, demenames = true)

Generates a String containing a Newick-format description of the Genealogy g. If extended = true (the default), then the PhyloPOMP-flavored Newick extension is used (i.e., metadata tags are included). If demenames = false, then demes are indicated using integers instead of names (in the extended format only).

PhyloPOMP.guide — Function

guide(g, m, knowledge!)

g is a Genealogy.

m: is an FSMarkovProc encoding the guiding finite-state Markov process. Construct it with a call to fsmarkov.

knowledge! is a function with signature knowledge!(v; deme, type, time). It should overwrite the vector v with probabilities of being in each possible deme, according to the given deme, type, and time. By default, knowledge! = known_deme! (see known_deme!).

PhyloPOMP.fsmarkov — Function

fsmarkov(...)

Constructs a finite-state Markov process in continuous time, represented as an object of type FSMarkovProc. Each argument is of the form D => prob or (D1,D2) => cond, where D, D1, D2 are demes, prob is a probability, and cond is a conductance. Specifically,

D => prob indicates that the stationary distribution on deme D is proportional to prob.

(D1,D2) => cond mandates that the conductance between D1 and D2 is cond.

PhyloPOMP.rcateg — Function

rcateg(p, prob = false)

If p is a vector of weights, rcateg(p) returns a draw from the categorical distribution on 1:length(p) with these weights. It also returns the sum of the weights.

If prob=true, the normalized weight of the selected category is returned as a third value.

It is not necessary for the weights to be normalized: this is accomplished internally.

rcateg(p, e, prob = false)

This call returns a draw from the categorical distribution on the enumeration e. e should be an enumeration-type created, e.g., by @demes.

rcateg(p, s, prob = false)

This call returns a draw from the categorical distribution on the set s. The latter may be a Set or a BitSet.

rcateg(p, v, prob = false)

This call returns a draw from the categorical distribution on the vector v.

PhyloPOMP.swap! — Function

swap!(y, i, j, b)

Swap lineage b from deme i to deme j in coloring y.

PhyloPOMP.chop! — Function

chop!(y, i, b)

Chop lineage b from deme i.

chop!(y, i, b, j, c)

Chop lineage b from deme i; add lineage c in deme j.

PhyloPOMP.fork! — Function

fork!(y, i, b, j, c)

Fork lineage b in deme i to lineages given in the vector c in demes given respectively by the vector j.