PhyloPOMP.jl

For doing phylodynamics in Julia.... It enables computation of the phylodynamic likelihood using tools from POMP.jl.

Function Documentation

PhyloPOMP.parse_newick — Function

parse_newick(input; demes, t0, time)

Parse the Newick-format string (or vector of strings) input.

Arguments:

demes is the enumeration of the demes.

t0 is the assumed root-time.

time is the (optional) final-time.

PhyloPOMP.newick — Function

newick(g; extended = true, sigdigits = 6)

Generates a Newick-format description of the Genealogy g.

Reference

PhyloPOMP.cap_tips! — Method

cap_tips!(G)

Converts tip-nodes to sample-nodes. The genealogy remains correct.

PhyloPOMP.newick — Method

newick(g; extended = true, sigdigits = 6)

Generates a Newick-format description of the Genealogy g.

PhyloPOMP.parse_newick — Method

parse_newick(input; demes, t0, time)

Parse the Newick-format string (or vector of strings) input.

Arguments:

demes is the enumeration of the demes.

t0 is the assumed root-time.

time is the (optional) final-time.

PhyloPOMP.repair! — Method

repair!(G)

Re-sorts the genealogical nodes. Renames all nodes and corrects the parent/child relationships.

PhyloPOMP.scan_branch! — Method

scan_branch!(input, G, p, mapper)

Parse the branch-string in input, appending the corresponding node to Genealogy G.

Arguments:

input: the string, or vector of strings, containing the branch information

G: the genealogy to be modified

p: the name of the parent node

mapper: a Dict mapping the deme-specification string to the correct deme.

PhyloPOMP.GenealNode — Type

GenealNode{D}

Implements a genealogical node. The Enum type D enumerates the demes.

PhyloPOMP.Genealogy — Type

Genealogy{D}

Implements a genealogy over a time interval [t0,t]. The Enum-type D enumerates the demes. Internally, this is represented as a time-ordered sequence of genealogical nodes (represented by GenealNode objects).